Split benches from package root to compile on stable
